Rename package common_state_context -> handler_setup
It reflects the package's purpose better as this
one is responsible for the common functions used
to setup task handlers throughout managers.
